+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- An `onProgress` handler option to attachment upload and submission functions
+
 ## [2.1.0] - 2022-09-13
 
 ### Added